Add legacy and modern build options for different React versions
This update adds separate build options for legacy and modern versions of React, allowing to switch between them as needed. It also enhances the descriptions for each project to clarify its distinct function and utilized technology. This change will help maintain and support apps and components built on different versions of React more effectively and efficiently.