Update package.json and README.md to modify start scripts
The start script in the package.json of the complete-react-case component app has been edited and a new legacy:start script has been added. Additionally, installation and start instructions in the README.md have been updated to use pnpm instead of npm. These changes seem to be focused on improving the build process and updating the project's documentation.