Update webpack loaders and add package description
Replaced Babel loader with swc-loader in rspack.config.js to better handle TypeScript and JSX syntax. A description was added to the package.json file for the "dynamic-remotes-node" package to provide a brief explanation of its functionality. The swc-loader was chosen for its speed and efficiency when transforming TypeScript and JSX code, and the package description will offer clearer context for users of the "dynamic-remotes-node" package.