Refactor build process and dynamic loading using rspack
In this commit, the build and serve process of the applications is overhauled to use rspack as the main tool instead of webpack-cli. The relevant scripts in package.json files of the three apps are modified accordingly. Additionally, functionality is added to perform dynamic remote loading of modules, in a manner that avoids redundant loading of modules that have already been pulled in once using module-federation runtime.