Implement rspack in various component npm scripts
The commit updates federated components to run and build with the 'rspack' tool in addition to webpack. The dependency '@rspack/core', '@rspack/cli', and '@rspack/dev-server' are added in each component's package.json, and the npm scripts 'start' and 'build' are adjusted to use 'rspack'. This is meant to extend the configuration capabilities beyond what 'webpack' offers. Meanwhile, for backward compatibility, legacy 'webpack' scripts are kept under 'legacy:start' and 'legacy:build'.