Refine package descriptions in multiple projects
Updated the descriptions in the 'package.json' files for multiple projects to be more concise and straight-to-the-point. These edited descriptions still cover the key details such as the tools and techniques used but in a more succinct manner. This should make it easier for users to quickly understand the core purpose of each project.