Remove unnecessary files and configurations
Deleted a substantial amount of files including app.e2e-spec.ts, app.po.ts, protractor.conf.js, tsconfig.json and a variety of profiles from angular12-microfrontends. This is part of a broader cleanup effort to streamline the project structure, remove redundancies and improve the maintainability of the source code.