Update script sleep time and clean up pnpm-lock.yaml
Updated the sleep time in the 'e2e:ci' script in package.json to improve execution time. Many entries in the pnpm-lock.yaml, notably for dynamic-system-host dependencies, were removed for better readability. Added functionality to the 'getDirectoryTree' function to recursively search for 'rspack' or 'webpack' configuration files within each directory. This results in a markdown file that now indicates if 'rspack' or 'webpack' are being used in a directory, providing more clarity in the directory structure.