Update "@rspack/core" and "@rspack/dev-server" versions (#3512)
* Update "@rspack/core" and "@rspack/dev-server" versions

Updated the versions of "@rspack/core" and "@rspack/dev-server" in several package.json files. The new version improves performance and provides additional features. These updates keep the project up-to-date with the latest development practices.

* update rspack

* Add local webpack and setup rspack for development in NextJs and React apps

This commit introduces rspack in the build and dev scripts of NextJs and React apps, replacing webpack-cli. We've added the NEXT_PRIVATE_LOCAL_WEBPACK environment variable in our build and dev scripts for enabling local webpack. In the React app, a new rspack.config.js file has been created for configuring the newly introduced rspack.

* locks

* Update consumers-react compatibility status

Changed the status of compatibility for consumers-react with rspack from not compatible to compatible. This update in the documentation reflects the recent updates and improvements made in the compatibility of the consumers-react with rspack.

* Update consumers-react compatibility status

Changed the status of compatibility for consumers-react with rspack from not compatible to compatible. This update in the documentation reflects the recent updates and improvements made in the compatibility of the consumers-react with rspack.

* Add extensions resolution to rspack.config.js

Added resolve property with extensions ['.ts', '.tsx', '.js'] to the configuration files of both app1 and app2. This allows for the resolution of TypeScript files and JavaScript files in the rspack module bundler, enhancing compatibility and flexibility.

* Update RemoteButton import path in Components.tsx

Changed the import path for RemoteButton in Components.tsx. Instead of 'app2/NoLongerWorks', the code now imports from 'app2/Button', ensuring efficient loading and operation of the Button component in the application.

* l

* locks

---------

Co-authored-by: ScriptedAlchemy <zackaryjackson@bytedance.com>