Migrate more examples (#3518)
* Update "@rspack/core" and "@rspack/dev-server" versions

Updated the versions of "@rspack/core" and "@rspack/dev-server" in several package.json files. The new version improves performance and provides additional features. These updates keep the project up-to-date with the latest development practices.

* update rspack

* Add local webpack and setup rspack for development in NextJs and React apps

This commit introduces rspack in the build and dev scripts of NextJs and React apps, replacing webpack-cli. We've added the NEXT_PRIVATE_LOCAL_WEBPACK environment variable in our build and dev scripts for enabling local webpack. In the React app, a new rspack.config.js file has been created for configuring the newly introduced rspack.

* locks

* chore(deps): update dependency eslint-plugin-import to v2.29.1

* Add resolve extensions in rspack.config.js

The update enables resolution of module paths for file types specified, which includes ".tsx", ".ts", ".jsx", ".js", ".json". This change will improve module dependency resolution by specifying the permissible file type extensions.

* Add resolve extensions to webpack configuration

Added ".tsx", ".ts", ".jsx", ".js", and ".json" extensions in the resolve array of the rspack.config.js in both app1 and app2. These additional extensions will aid in the resolution of module paths, consequently enhancing module dependency management.

* Update webpack and rspack configurations

Modified webpack.config.js in app-02 by removing some unnecessary spaces and included ".jsx", ".js", ".json", and ".mjs" extensions in the resolve configurations of the rspack.config.js in app-01, app-02 and app-03. This is to enhance the resolution of module paths and to improve module dependency management.

* chore(deps): update dependency less-loader to v11.1.4

* Update .gitignore with new paths

Added new entries to .gitignore for JS and CSS files in the comprehensive-demo-react16 public folder. This change prevents unneeded static build files from being tracked and committed to the repository.

* Update .gitignore with new paths

Added new entries to .gitignore for JS and CSS files in the comprehensive-demo-react16 public folder. This change prevents unneeded static build files from being tracked and committed to the repository.

* Update consumers-react compatibility status

Changed the status of compatibility for consumers-react with rspack from not compatible to compatible. This update in the documentation reflects the recent updates and improvements made in the compatibility of the consumers-react with rspack.

* Update consumers-react compatibility status

Changed the status of compatibility for consumers-react with rspack from not compatible to compatible. This update in the documentation reflects the recent updates and improvements made in the compatibility of the consumers-react with rspack.

* chore(deps): update dependency postcss to v8.4.33

* locks

* locks

---------

Co-authored-by: ScriptedAlchemy <zackaryjackson@bytedance.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>