Update scripts and dependencies in package.json files
In this commit, the scripts and dependencies within the package.json files of multiple applications and the remote library have been updated. The Webpack commands were replaced by rspack for the 'start' and 'build' scripts, while preserving the former as legacy options. New rspack related devDependencies were also added. The scripts on the root level package.json were updated to use pnpm, optimizing for a monorepo setup. Finally, the 'dev' script in the remoteLibrary was renamed to 'start'.