Update Next.js config and add new dependency
Updated the build and dev scripts in the package.json files of the host and remote to include NEXT_PRIVATE_LOCAL_WEBPACK, providing local webpack configuration. Also, updated the NextFederationPlugin in the next.config file and removed the skipSharingNextInternals option. Added the "wait-on" dependency in the devDependencies of the package.json file.