Upgrade rsbuild (#3548)
* Update rsbuild core and plugin versions

This commit updates the rsbuild core and plugin versions in the devDependencies sections of packages.json files for cra/remote, cra/host, and angular15-vue3/vue-remote. The changes will make sure we're using the latest versions of these dependencies.

* Update rsbuild core and plugin versions

This commit updates the rsbuild core and plugin versions in the devDependencies sections of packages.json files for cra/remote, cra/host, and angular15-vue3/vue-remote. The changes will make sure we're using the latest versions of these dependencies.

* Update rsbuild core and plugin versions

This commit updates the rsbuild core and plugin versions in the devDependencies sections of packages.json files for cra/remote, cra/host, and angular15-vue3/vue-remote. The changes will make sure we're using the latest versions of these dependencies.

* Update rsbuild core and plugin versions

This commit updates the rsbuild core and plugin versions in the devDependencies sections of packages.json files for cra/remote, cra/host, and angular15-vue3/vue-remote. The changes will make sure we're using the latest versions of these dependencies.

---------

Co-authored-by: ScriptedAlchemy <zackaryjackson@bytedance.com>