Adding new remote-router and corresponding host with webpack configurations
This commit introduces a new remote-router in the runtime-plugins directory and its corresponding host which includes its own set of webpack configuration files. It also includes creation, configuration, and inclusion of new JavaScript and TypeScript files for the new remote-router. Another important change is the update to package versions and dependencies within the pnpm-lock.yaml file.