feat: remote ingress runtime plugin (#3613)
* Adding new remote-router and corresponding host with webpack configurations

This commit introduces a new remote-router in the runtime-plugins directory and its corresponding host which includes its own set of webpack configuration files. It also includes creation, configuration, and inclusion of new JavaScript and TypeScript files for the new remote-router. Another important change is the update to package versions and dependencies within the pnpm-lock.yaml file.

* Add "e2e:ci" script in remote-router's package.json file

This commit adds a new script called "e2e:ci" in the package.json file for remote-router. It's a placeholder command echoing 'no tests' for now, meant to indicate where end-to-end tests for continuous integration would eventually go.

* Update dependencies and add "e2e:ci" script in package.json files

The commit introduces a modification to the dependencies by switching the associated version of html-webpack-plugin to the rspack core instead of webpack. It also adds a "e2e:ci" script to both host and remotes-monorepo package.json files, serving as a placeholder for future end-to-end tests.

* Downgrade eslint version across multiple dependencies

This changes updates various dependencies in the pnpm-lock.yaml file, primarily focusing on downgrading the eslint version from 8.56.0 to 8.48.0. These modifications ensure consistent linting standards

* Add README for remote-router runtime plugin

A new README.md has been created for the remote-router runtime plugin, detailing its role, functionality, and methods in the microfrontend architecture. This document provides key insights into how the plugin handles the loading and error management for remote microfrontends and aims to enhance understanding for future development and maintenance of the system.

* Improve remote-router plugin and add documentation

This commit enhances the remote-router runtime plugin by condensing code for error handling and remote entry management. A detailed comment section was added as documentation at the start of the file to clarify the plugin's functionality, how it works, and its significance within the module federation micro-frontend architecture.